requirements.txt
djoser==2.0.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==3.3.2
defusedxml==0.6.0
Django==2.2.24
django-crispy-forms==1.9.0
djangorestframework==3.11.2
urllib3==1.25.9
idna==2.9
mysqlclient==1.4.6
oauthlib==3.1.0
pycparser==2.20
Pillow==8.0.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.0
urllib3==1.25.9
urllib3==1.26.5
